+++ src/queryproc/summarizer/summarizerAccumulateVariable.cpp
@@ -53,7 +53,7 @@ void SummarizerFunctionContextAccumulateVariable::addSummarizationFeature(
 		const std::string& name,
 		PostingIteratorInterface* itr,
 		const std::vector<SummarizationVariable>& variables,
-		float weight,
+		double weight,
 		const TermStatistics&)
 {
 	try
