+++ src/queryproc/summarizer/summarizerAttribute.cpp
@@ -32,7 +32,7 @@ void SummarizerFunctionContextAttribute::addSummarizationFeature(
 		const std::string&,
 		PostingIteratorInterface*,
 		const std::vector<SummarizationVariable>&,
-		float /*weight*/,
+		double /*weight*/,
 		const TermStatistics&)
 {
 	m_errorhnd->report( _TXT( "no sumarization features expected in summarization function '%s'"), "attribute");
