+++ src/queryproc/summarizer/summarizerListMatches.cpp
@@ -25,7 +25,7 @@ void SummarizerFunctionContextListMatches::addSummarizationFeature(
 		const std::string& name,
 		PostingIteratorInterface* itr,
 		const std::vector<SummarizationVariable>&,
-		float /*weight*/,
+		double /*weight*/,
 		const TermStatistics&)
 {
 	try
