+++ src/queryproc/summarizer/summarizerMatchPhrase.hpp
@@ -76,7 +76,7 @@ public:
 			const std::string& name,
 			PostingIteratorInterface* itr,
 			const std::vector<SummarizationVariable>&,
-			float weight,
+			double weight,
 			const TermStatistics&);
 
 	virtual std::vector<SummaryElement> getSummary( const Index& docno);
