+++ src/queryproc/summarizer/summarizerMatchVariables.hpp
@@ -54,7 +54,7 @@ public:
 			const std::string& name,
 			PostingIteratorInterface* itr,
 			const std::vector<SummarizationVariable>& variables,
-			float weight,
+			double weight,
 			const TermStatistics&);
 
 	virtual std::vector<SummaryElement> getSummary( const Index& docno);
@@ -64,9 +64,9 @@ private:
 	{
 		PostingIteratorInterface* itr;
 		std::vector<SummarizationVariable> variables;
-		float weight;
+		double weight;
 
-		SummarizationFeature( PostingIteratorInterface* itr_, const std::vector<SummarizationVariable>& variables_, float weight_)
+		SummarizationFeature( PostingIteratorInterface* itr_, const std::vector<SummarizationVariable>& variables_, double weight_)
 			:itr(itr_),variables(variables_),weight(weight_){}
 		SummarizationFeature( const SummarizationFeature& o)
 			:itr(o.itr),variables(o.variables),weight(o.weight){}
