+++ src/queryproc/summarizer/summarizerMetaData.hpp
@@ -47,7 +47,7 @@ public:
 			const std::string&,
 			PostingIteratorInterface*,
 			const std::vector<SummarizationVariable>&,
-			float /*weight*/,
+			double /*weight*/,
 			const TermStatistics&);
 
 	virtual std::vector<SummaryElement> getSummary( const Index& docno);
