+++ src/queryproc/weighting/weightingFrequency.hpp
@@ -34,9 +34,9 @@ public:
 	struct Feature
 	{
 		PostingIteratorInterface* itr;
-		float weight;
+		double weight;
 
-		Feature( PostingIteratorInterface* itr_, float weight_)
+		Feature( PostingIteratorInterface* itr_, double weight_)
 			:itr(itr_),weight(weight_){}
 		Feature( const Feature& o)
 			:itr(o.itr),weight(o.weight){}
@@ -45,7 +45,7 @@ public:
 	virtual void addWeightingFeature(
 			const std::string& name_,
 			PostingIteratorInterface* itr_,
-			float weight_,
+			double weight_,
 			const TermStatistics&);
 
 	virtual double call( const Index& docno);
