+++ src/queryproc/weighting/weightingSmart.cpp
@@ -46,7 +46,7 @@ WeightingFunctionContextSmart::WeightingFunctionContextSmart(
 void WeightingFunctionContextSmart::addWeightingFeature(
 		const std::string& name_,
 		PostingIteratorInterface* itr_,
-		float weight_,
+		double weight_,
 		const TermStatistics& stats_)
 {
 	try
